refactor: segregate outdent handler
